all tasks that i could this moment

+++ src/by/it/drachov/lesson02/TaskA2.java
@@ -10,7 +10,7 @@ package by.it.drachov.lesson02;
 
 class TaskA2 {
     public static void main(String[] args) {
-        for (int i = 0; i < 5; i++) {
+        for (byte i = 0; i < 5; i++) {
             System.out.println("Я начинаю изучать Java!");
 
         }
